Package re-organization for setup.py
