+++ c3d/__init__.py
@@ -1,0 +1,8 @@
+from . import dtypes
+from . import group
+from . import header
+from . import manager
+from . import parameter
+from . import utils
+from .reader import Reader
+from .writer import Writer
