+++ setup.py
@@ -4,7 +4,7 @@ import setuptools
 setuptools.setup(
     name='c3d',
     version='0.3.0',
-    py_modules=['c3d'],
+    py_modules=['c3d.dtypes', 'c3d.group', 'c3d.header', 'c3d.manager', 'c3d.reader', 'c3d.writer', 'c3d.parameter', 'c3d.utils', 'scripts.c3d-viewer'],
     author='UT Vision, Cognition, and Action Lab',
     author_email='leif@cs.utexas.edu',
     description='A library for manipulating C3D binary files',
@@ -13,7 +13,7 @@ setuptools.setup(
     url='http://github.com/EmbodiedCognition/py-c3d',
     keywords=('c3d motion-capture'),
     install_requires=['numpy'],
-    scripts=['scripts/c3d{}'.format(s) for s in '-metadata -viewer 2csv 2npz'.split()],
+    scripts=['scripts/c3d{}.py'.format(s) for s in '-metadata -viewer 2csv 2npz'.split()],
     classifiers=[
         'Development Status :: 4 - Beta',
         'Intended Audience :: Science/Research',
